Share the FFI string reader between PropertyGrid and ComboBox
