+++ rust/wxdragon/src/utils/mod.rs
@@ -1,7 +1,9 @@
 mod array_string;
 mod colour_database;
+mod ffi_string;
 mod misc;
 
 pub use array_string::ArrayString;
 pub use colour_database::{ColourDatabase, ColourDatabaseScheme};
+pub(crate) use ffi_string::read_ffi_string;
 pub use misc::{BrowserLaunchFlags, bell, get_key_state, get_mouse_position, launch_default_application, launch_default_browser};
